requirements.txt
aiofiles
aiohttp
discord.py
googletrans
openai
pillow
psutil
py-cpuinfo
pytesseract
PyYAML
serpapi
yt-dlp
google_search_results